\name{NumericLogical-class}
\Rdversion{1.1}
\docType{class}
\alias{NumericLogical-class}

\title{Class "NumericLogical"  }
\description{
    An S4 class union, whose member classes are class \code{"numeric"} and class \code{"logical"}.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\section{Methods}{
No methods defined with class "NumericLogical" in the signature.
}

\examples{
showClass("NumericLogical")
}

\keyword{classes}
